+++ web/implement/ArkWebController.php
@@ -14,36 +14,54 @@ use sinri\ark\io\ArkWebOutput;
 class ArkWebController
 {
 
-    protected $shouldSendJsonHeader = true;
     protected $filterGeneratedData;
 
+    /**
+     * ArkWebController constructor.
+     * You should process filters-generated-data here if needed,
+     * the property $filterGeneratedData is set here.
+     */
     public function __construct()
     {
-        // If you want to track a request, use
-        // $request_serial = Ark()->webService()->getRequestSerial();
-
-        // You might process filters-generated-data here
         $this->filterGeneratedData = Ark()->webService()->getFilterGeneratedData();
     }
 
+    /**
+     * If you want to track a request, use this method to get the serial number of a request.
+     * @return string
+     */
+    protected function _getRequestSerial()
+    {
+        return Ark()->webService()->getRequestSerial();
+    }
+
+    /**
+     * @param string $data
+     * @param int $httpCode
+     */
     protected function _sayOK($data = "", $httpCode = 200)
     {
-        if ($this->shouldSendJsonHeader) {
-            Ark()->webOutput()->setContentTypeHeader(ArkWebOutput::CONTENT_TYPE_JSON);
-        }
+        Ark()->webOutput()->setContentTypeHeader(ArkWebOutput::CONTENT_TYPE_JSON);
         http_response_code($httpCode);
         Ark()->webOutput()->jsonForAjax(ArkWebOutput::AJAX_JSON_CODE_OK, $data);
     }
 
+    /**
+     * @param string $error
+     * @param int $httpCode
+     */
     protected function _sayFail($error = "", $httpCode = 200)
     {
-        if ($this->shouldSendJsonHeader) {
-            Ark()->webOutput()->setContentTypeHeader(ArkWebOutput::CONTENT_TYPE_JSON);
-        }
+        Ark()->webOutput()->setContentTypeHeader(ArkWebOutput::CONTENT_TYPE_JSON);
         http_response_code($httpCode);
         Ark()->webOutput()->jsonForAjax(ArkWebOutput::AJAX_JSON_CODE_FAIL, $error);
     }
 
+    /**
+     * @param string $templateFile
+     * @param array $params
+     * @param int $httpCode
+     */
     protected function _showPage($templateFile, $params = [], $httpCode = 200)
     {
         http_response_code($httpCode);