+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### pre Version 0.9
 
 Remove 'force logging to STDOUT' feature in ArkLogger and designed `echo` instead.
+`ArkPDO` treat database config as optional.
+`ArkCurl` support setting raw options of CURL.
 
 ### Version 0.8